requirements.txt
annotated-types==0.7.0 ; python_full_version == "3.11.4"
anyio==4.8.0 ; python_full_version == "3.11.4"
blinker==1.9.0 ; python_full_version == "3.11.4"
certifi==2024.12.14 ; python_full_version == "3.11.4"
click==8.1.8 ; python_full_version == "3.11.4"
colorama==0.4.6 ; python_full_version == "3.11.4" and platform_system == "Windows"
distro==1.9.0 ; python_full_version == "3.11.4"
flask-cors==5.0.0 ; python_full_version == "3.11.4"
flask==3.1.0 ; python_full_version == "3.11.4"
gunicorn==23.0.0 ; python_full_version == "3.11.4"
h11==0.14.0 ; python_full_version == "3.11.4"
httpcore==1.0.7 ; python_full_version == "3.11.4"
httpx==0.28.1 ; python_full_version == "3.11.4"
idna==3.10 ; python_full_version == "3.11.4"
itsdangerous==2.2.0 ; python_full_version == "3.11.4"
jinja2==3.1.5 ; python_full_version == "3.11.4"
jiter==0.8.2 ; python_full_version == "3.11.4"
markupsafe==3.0.2 ; python_full_version == "3.11.4"
openai==1.59.6 ; python_full_version == "3.11.4"
packaging==24.2 ; python_full_version == "3.11.4"
pydantic-core==2.27.2 ; python_full_version == "3.11.4"
pydantic==2.10.5 ; python_full_version == "3.11.4"
python-dotenv==1.0.1 ; python_full_version == "3.11.4"
sniffio==1.3.1 ; python_full_version == "3.11.4"
tqdm==4.67.1 ; python_full_version == "3.11.4"
typing-extensions==4.12.2 ; python_full_version == "3.11.4"
werkzeug==3.1.3 ; python_full_version == "3.11.4"